Fixed a few projection bugs - missing delegate returns, ppl.h generation, AsyncActionProgressHandler traits, IReference ctor, duplicate begin/end defs...

Restored IReference production support (this depends on W.F.h, now #included in base.h, and needs to be refactored)

Restored std::hash implementations for interfaces, classes

With relocation of guids from abi types to classes, it's no longer possible to reach the guid from either artifact.  The result is that it's no longer possible to mix and match the two.  Obsoleted test code in com_ptr.cpp, IUnknown_QI.cpp has been deleted.

IReferenced required some category goo for my_enum and my_struct - not sure if it's correct.

Removed generic guids (yeah!)

Created test certificates for a few samples to get them building

+++ cppwinrt/strings/base/base_string.h
@@ -23,7 +23,7 @@ namespace impl
             length = wcslen(value);
         }
         WINRT_ASSERT(length <= 0xffffffff);
-        return length;
+        return static_cast<uint32_t>(length);
     }
 
     inline bool embedded_null(HSTRING value) noexcept
